+++ dtaianomaly/anomaly_detection/__init__.py
@@ -14,6 +14,7 @@ from .windowing_utils import sliding_window, reverse_sliding_window, check_is_va
 from .baselines import AlwaysNormal, AlwaysAnomalous, RandomDetector
 from .PyODAnomalyDetector import PyODAnomalyDetector
 from .ClusterBasedLocalOutlierFactor import ClusterBasedLocalOutlierFactor
+from .CopulaBasedOutlierDetector import CopulaBasedOutlierDetector
 from .HistogramBasedOutlierScore import HistogramBasedOutlierScore
 from .IsolationForest import IsolationForest
 from .KernelPrincipalComponentAnalysis import KernelPrincipalComponentAnalysis
@@ -45,6 +46,7 @@ __all__ = [
 
     # Detectors
     'ClusterBasedLocalOutlierFactor',
+    'CopulaBasedOutlierDetector',
     'HistogramBasedOutlierScore',
     'IsolationForest',
     'KernelPrincipalComponentAnalysis',